Make build insist on JDK 1.8.

+++ build/javadocs.sh
@@ -1,8 +1,12 @@
 #!/bin/sh
+if [ "$JAVA8_HOME" = "" ] ; then
+    echo "Please set the JAVA8_HOME environment variable."
+    exit 1
+fi
 cd `dirname $0`/..
 echo "Running in `pwd`"
 rm -rf docs/javadocs
-javadoc -d docs/javadocs -sourcepath src edu.calpoly.testy
+$JAVA8_HOME/bin/javadoc -d docs/javadocs -sourcepath src edu.calpoly.testy
 cd docs
 cp -r javadocs testy-javadocs
 mkdir -p ../out
